Minor progress, hexadecimals still not working

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: rnovotny <rnovotny@student.42prague.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/28 15:38:54 by rnovotny          #+#    #+#             */
-/*   Updated: 2023/02/19 20:34:37 by rnovotny         ###   ########.fr       */
+/*   Updated: 2023/03/03 21:10:37 by rnovotny         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int		ft_uint(va_list ap, char c);
 int		ft_hex(va_list ap, char c);
 
 int		ft_puthex(unsigned long long n, int neg, char c);
-int		ft_putnbr(long n, int neg);
+int		ft_putnbr(unsigned int n, int neg);
 int		ft_put(long n, char c);
 size_t	ft_strlen(const char *str);
